// src/sdk/generated/common/types/getAllPatientsParams.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export type GetAllPatientsParams = {
  search?: string;
  gender?: string;
  age?: number;
  /**
   * @minimum 0
   */
  page?: number;
  /**
   * @minimum 1
   */
  size?: number;
  sort?: string[];
};